+++ CHANGELOG.md
@@ -11,6 +11,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
   * Pricing didn't update when target mint supply changed
 ## [2.0.6](https://github.com/ChewingGlassFund/wumbo-programs/compare/v2.0.5...v2.0.6) (2021-12-30)
   * Fix account fetch cache to output parsed data when a parser is provided for something cached
+  * Add more options to bonding parameters to social token
 
 
 ## [2.0.5](https://github.com/ChewingGlassFund/wumbo-programs/compare/v2.0.4...v2.0.5) (2021-12-30)